# docs/openapi/swagger.json
{
  "components": {
    "schemas": {
...
      "BatchUpdateInquiryStatusRequest": {
        "properties": {
          "excludedInquiryIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/FilterCriteria"
          },
          "inquiryIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "mode": {
            "enum": [
              "IDS",
              "FILTER"
            ],
            "type": "string"
          },
          "reason": {
            "maxLength": 2147483647,
            "minLength": 5,
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED"
            ],
            "type": "string"
          }
        },
        "required": [
          "mode",
          "reason",
          "status"
        ],
        "type": "object"
      },
...
      "ChannelMetadata": {
        "discriminator": {
          "propertyName": "metadataType"
        },
        "properties": {
          "metadataType": {
            "type": "string"
          }
        },
        "required": [
          "metadataType"
        ],
        "type": "object"
      },
      "Content": {
        "properties": {
          "channel": {
            "type": "string"
          },
          "channelMetadata": {
            "$ref": "#/components/schemas/ChannelMetadata"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deviceInfo": {
            "$ref": "#/components/schemas/DeviceInfo"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "imageUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "isManual": {
            "type": "boolean"
          },
          "parentId": {
            "format": "uuid",
            "type": "string"
          },
          "replyCount": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "uniqueKey": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userCode": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateInquiryRequest": {
        "properties": {
          "channel": {
            "type": "string"
          },
          "channelMetadata": {
            "$ref": "#/components/schemas/ChannelMetadata"
          },
          "content": {
            "type": "string"
          },
          "deviceInfo": {
            "$ref": "#/components/schemas/DeviceInfo"
          },
          "imageUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "userCode": {
            "type": "string"
          }
        },
        "required": [
          "channel",
          "content"
        ],
        "type": "object"
      },
...
      "DeviceInfo": {
        "properties": {
          "appVersion": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "osVersion": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "FilterCriteria": {
        "properties": {
          "bookmarkedOnly": {
            "type": "boolean"
          },
          "channel": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "end": {
            "format": "date-time",
            "type": "string"
          },
          "isManual": {
            "type": "boolean"
          },
          "start": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "items": {
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "RESOLVED"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "userCode": {
            "type": "string"
          },
          "userCodeMissing": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "InquiryCountResponse": {
        "properties": {
          "count": {
            "format": "int64",
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "OperatorInfo": {
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SearchCustomerInquiryResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/Content"
            },
            "type": "array"
          },
          "hasNext": {
            "type": "boolean"
          },
          "nextCursor": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateInquiryRequest": {
        "properties": {
          "channel": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "customFields": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object"
          },
          "deviceInfo": {
            "$ref": "#/components/schemas/DeviceInfo"
          },
          "imageUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "operatorInfo": {
            "$ref": "#/components/schemas/OperatorInfo"
          },
          "reasons": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "status": {
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED"
            ],
            "type": "string"
          },
          "userCode": {
            "type": "string"
          }
        },
        "required": [
          "operatorInfo"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "InternalToken": {
        "description": "내부 시스템 간의 보안 통신을 위한 인증 토큰입니다.",
        "in": "header",
        "name": "X-Internal-Token",
        "type": "apiKey"
      },
      "NginxBasicAuth": {
        "description": "nginx Basic Auth 인증입니다. 인증 성공 후 nginx가 X-Remote-User 헤더를 backend로 전달합니다.",
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "고객 서비스 백엔드 API 명세서입니다. 문의 사항 관리, 웹훅 처리, 네이버 세션 관리 등의 기능을 제공합니다.\n\n/api/** endpoint는 nginx Basic Auth 인증 후 전달되는 X-Remote-User 또는 내부 시스템용 X-Internal-Token 인증이 필요합니다.\n@RequireRoles가 적용된 endpoint는 문서에 필요 역할과 x-required-roles 확장값으로 표시됩니다.\n",
    "title": "CS Test Bed API 명세서",
    "version": "1.0.0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/inquiries": {
      "get": {
        "description": "채널, 유저 코드, 상태, 시작/종료 시간 및 커서를 조합하여 고객 문의 내역 목록을 조회합니다.",
        "operationId": "search",
        "parameters": [
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "userCode",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "enum": [
                  "OPEN",
                  "IN_PROGRESS",
                  "RESOLVED"
                ],
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isManual",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "bookmarkedOnly",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "userCodeMissing",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCustomerInquiryResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "고객 문의 내역 검색 및 조회",
        "tags": [
          "Customer Inquiry API"
        ]
      },
      "post": {
        "description": "신규 고객 문의 건을 접수 및 생성합니다.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInquiryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "고객 문의 생성",
        "tags": [
          "Customer Inquiry API"
        ]
      }
    },
    "/api/v1/inquiries/batch/status": {
      "patch": {
        "description": "여러 고객 문의의 진행 상태를 일괄 수정하고 각각 이력을 남깁니다.",
        "operationId": "updateStatuses",
        "parameters": [
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdateInquiryStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "고객 문의 일괄 상태 수정",
        "tags": [
          "Customer Inquiry API"
        ]
      }
    },
...
    "/api/v1/inquiries/count": {
      "get": {
        "description": "목록 데이터를 내려주지 않고 조건에 맞는 문의 수와 초과 여부만 조회합니다.",
        "operationId": "count",
        "parameters": [
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "userCode",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "enum": [
                  "OPEN",
                  "IN_PROGRESS",
                  "RESOLVED"
                ],
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isManual",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "bookmarkedOnly",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "userCodeMissing",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryCountResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "고객 문의 내역 조건별 카운트 조회",
        "tags": [
          "Customer Inquiry API"
        ]
      }
    },
...
    "/api/v1/inquiries/{id}": {
      "patch": {
        "description": "네이버 카페 채널의 경우, 해당 카페 게시글의 상세 정보 및 신규 댓글들을 네이버 API를 통해 즉시 동기화합니다.",
        "operationId": "updateInquiry_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInquiryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Content"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "고객 문의 정보 최신화",
        "tags": [
          "Customer Inquiry API"
        ]
      }
    },
    "/api/v1/inquiries/{id}/replies": {
      "get": {
        "description": "특정 문의의 모든 자식 문의 목록을 시간 순으로 조회합니다.",
        "operationId": "getReplies",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Content"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "이메일 회신(자식 문의) 목록 조회",
        "tags": [
          "Customer Inquiry API"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://cs-api:8080"
    }
  ],
...
}